@@ -94,12 +95,12 @@
     "coveralls": "3.1.0",
     "del": "5.1.0",
     "dependency-graph": "0.9.0",
-    "eslint": "7.6.0",
+    "eslint": "7.7.0",
     "eslint-plugin-import": "2.22.0",
     "express": "4.17.1",
     "find-free-port": "2.0.0",
-    "firebase-functions": "3.9.0",
-    "firebase-tools": "8.6.0",
+    "firebase-functions": "3.11.0",
+    "firebase-tools": "8.9.0",
     "git-rev-sync": "2.0.0",
     "glob": "7.1.6",
     "http-server": "0.12.3",
