@@ -1,3 +1,20 @@
+/**
+ * @license
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 const path = require('path');
 
 module.exports = {
@@ -128,12 +145,29 @@ module.exports = {
           'Object': "Use {} or 'object' instead.",
           'String': "Use 'string' instead.",
           'Number': "Use 'number' instead.",
-          'Boolean': "Use 'boolean' instead."
+          'Boolean': "Use 'boolean' instead.",
+          'Function': `Avoid the Function type, as it provides little safety for the following reasons:
+                       It provides no type safety when calling the value, which means it's easy to provide the wrong arguments.
+                       It accepts class declarations, which will fail when called, as they are called without the new keyword.`
+        },
+        'extendDefaults': false
+      }
+    ],
+    '@typescript-eslint/naming-convention': [
+      'error',
+      {
+        'selector': 'class',
+        'format': ['PascalCase']
+      },
+      {
+        'selector': 'interface',
+        'format': ['PascalCase'],
+        'custom': {
+          'regex': '^I[A-Z]',
+          'match': false
         }
       }
     ],
-    '@typescript-eslint/class-name-casing': 'error',
-    '@typescript-eslint/interface-name-prefix': ['error', 'never'],
     '@typescript-eslint/consistent-type-definitions': ['error', 'interface'],
     '@typescript-eslint/explicit-member-accessibility': [
       'error',
