@@ -56,7 +56,7 @@ export interface CompleteChildSource {
  * @constructor
  * @implements CompleteChildSource
  */
-// eslint-disable-next-line @typescript-eslint/class-name-casing
+// eslint-disable-next-line @typescript-eslint/naming-convention
 export class NoCompleteChildSource_ implements CompleteChildSource {
   /**
    * @inheritDoc
