@@ -15,12 +15,14 @@
  * limitations under the License.
  */
 
+import tmp from 'tmp';
 import json from 'rollup-plugin-json';
 import alias from '@rollup/plugin-alias';
 import typescriptPlugin from 'rollup-plugin-typescript2';
 import typescript from 'typescript';
 import path from 'path';
-import sourcemaps from 'rollup-plugin-sourcemaps';
+import replace from 'rollup-plugin-replace';
+import copy from 'rollup-plugin-copy-assets';
 import { terser } from 'rollup-plugin-terser';
 import { importPathTransformer } from '../../scripts/exp/ts-transform-import-path';
 
@@ -36,11 +38,18 @@ const nodePlugins = [
         target: 'es2017'
       }
     },
-    clean: true,
+    cacheDir: tmp.dirSync(),
     abortOnError: false,
     transformers: [util.removeAssertTransformer, importPathTransformer]
   }),
-  json({ preferConst: true })
+  json({ preferConst: true }),
+  // Needed as we also use the *.proto files
+  copy({
+    assets: ['./src/protos']
+  }),
+  replace({
+    'process.env.FIRESTORE_PROTO_ROOT': JSON.stringify('../src/protos')
+  })
 ];
 
 const browserPlugins = [
@@ -51,7 +60,7 @@ const browserPlugins = [
         target: 'es2017'
       }
     },
-    clean: true,
+    cacheDir: tmp.dirSync(),
     abortOnError: false,
     transformers: [
       util.removeAssertAndPrefixInternalTransformer,
@@ -86,18 +95,7 @@ const allBuilds = [
       name: 'firebase.firestore',
       sourcemap: true
     },
-    plugins: [
-      typescriptPlugin({
-        typescript,
-        compilerOptions: {
-          allowJs: true,
-          target: 'es5'
-        },
-        include: ['dist/exp/*.js']
-      }),
-      json(),
-      sourcemaps()
-    ],
+    plugins: util.es2017ToEs5Plugins(/* mangled= */ false),
     external: util.resolveNodeExterns,
     treeshake: {
       moduleSideEffects: false
