@@ -15,6 +15,7 @@
  * limitations under the License.
  */
 
+import tmp from 'tmp';
 import path from 'path';
 import json from 'rollup-plugin-json';
 import alias from '@rollup/plugin-alias';
@@ -36,7 +37,7 @@ const nodePlugins = [
         target: 'es2017'
       }
     },
-    clean: true,
+    cacheDir: tmp.dirSync(),
     abortOnError: false,
     transformers: [util.removeAssertTransformer, importPathTransformer]
   }),
@@ -51,7 +52,7 @@ const browserPlugins = [
         target: 'es2017'
       }
     },
-    clean: true,
+    cacheDir: tmp.dirSync(),
     abortOnError: false,
     transformers: [
       util.removeAssertAndPrefixInternalTransformer,
