@@ -15,9 +15,33 @@
  * limitations under the License.
  */
 
-import * as firestore from '@firebase/firestore-types';
-
-import * as api from '../protos/firestore_proto_api';
+import { Value as ProtoValue } from '../protos/firestore_proto_api';
+import {
+  CollectionReference as PublicCollectionReference,
+  DocumentChange as PublicDocumentChange,
+  DocumentChangeType,
+  DocumentData,
+  DocumentReference as PublicDocumentReference,
+  DocumentSnapshot as PublicDocumentSnapshot,
+  FirebaseFirestore as PublicFirestore,
+  FirestoreDataConverter,
+  GetOptions,
+  LogLevel as PublicLogLevel,
+  OrderByDirection,
+  PersistenceSettings as PublicPersistenceSettings,
+  Query as PublicQuery,
+  QueryDocumentSnapshot as PublicQueryDocumentSnapshot,
+  QuerySnapshot as PublicQuerySnapshot,
+  SetOptions,
+  Settings as PublicSettings,
+  SnapshotListenOptions,
+  SnapshotMetadata as PublicSnapshotMetadata,
+  SnapshotOptions as PublicSnapshotOptions,
+  Transaction as PublicTransaction,
+  UpdateData,
+  WhereFilterOp,
+  WriteBatch as PublicWriteBatch
+} from '@firebase/firestore-types';
 
 import { FirebaseApp } from '@firebase/app-types';
 import { _FirebaseApp, FirebaseService } from '@firebase/app-types/private';
@@ -130,7 +154,7 @@ export const CACHE_SIZE_UNLIMITED = LruParams.COLLECTION_DISABLED;
 const DEFAULT_SYNCHRONIZE_TABS = false;
 
 /** Undocumented, private additional settings not exposed in our public API. */
-interface PrivateSettings extends firestore.Settings {
+interface PrivateSettings extends PublicSettings {
   // Can be a google-auth-library or gapi client.
   credentials?: CredentialsSettings;
 }
@@ -160,7 +184,7 @@ class FirestoreSettings {
 
   readonly cacheSizeBytes: number;
 
-  readonly forceLongPolling: boolean;
+  readonly experimentalForceLongPolling: boolean;
 
   readonly ignoreUndefinedProperties: boolean;
 
@@ -263,7 +287,7 @@ class FirestoreSettings {
       'experimentalForceLongPolling',
       settings.experimentalForceLongPolling
     );
-    this.forceLongPolling =
+    this.experimentalForceLongPolling =
       settings.experimentalForceLongPolling ?? DEFAULT_FORCE_LONG_POLLING;
   }
 
@@ -274,7 +298,8 @@ class FirestoreSettings {
       this.timestampsInSnapshots === other.timestampsInSnapshots &&
       this.credentials === other.credentials &&
       this.cacheSizeBytes === other.cacheSizeBytes &&
-      this.forceLongPolling === other.forceLongPolling &&
+      this.experimentalForceLongPolling ===
+        other.experimentalForceLongPolling &&
       this.ignoreUndefinedProperties === other.ignoreUndefinedProperties
     );
   }
@@ -283,7 +308,7 @@ class FirestoreSettings {
 /**
  * The root reference to the database.
  */
-export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
+export class Firestore implements PublicFirestore, FirebaseService {
   // The objects that are a part of this API are exposed to third-parties as
   // compiled javascript so we want to flag our private members with a leading
   // underscore to discourage their use.
@@ -357,9 +382,15 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
     return this._userDataReader;
   }
 
-  settings(settingsLiteral: firestore.Settings): void {
+  settings(settingsLiteral: PublicSettings): void {
     validateExactNumberOfArgs('Firestore.settings', arguments, 1);
     validateArgType('Firestore.settings', 'object', 1, settingsLiteral);
+
+    if (settingsLiteral.merge) {
+      settingsLiteral = { ...this._settings, ...settingsLiteral };
+      // Remove the property from the settings once the merge is completed
+      delete settingsLiteral.merge;
+    }
 
     const newSettings = new FirestoreSettings(settingsLiteral);
     if (this._firestoreClient && !this._settings.isEqual(newSettings)) {
@@ -387,7 +418,7 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
     return this._firestoreClient!.disableNetwork();
   }
 
-  enablePersistence(settings?: firestore.PersistenceSettings): Promise<void> {
+  enablePersistence(settings?: PublicPersistenceSettings): Promise<void> {
     if (this._firestoreClient) {
       throw new FirestoreError(
         Code.FAILED_PRECONDITION,
@@ -516,7 +547,7 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
       this._persistenceKey,
       this._settings.host,
       this._settings.ssl,
-      this._settings.forceLongPolling
+      this._settings.experimentalForceLongPolling
     );
   }
 
@@ -582,7 +613,7 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
     }
   };
 
-  collection(pathString: string): firestore.CollectionReference {
+  collection(pathString: string): PublicCollectionReference {
     validateExactNumberOfArgs('Firestore.collection', arguments, 1);
     validateArgType('Firestore.collection', 'non-empty string', 1, pathString);
     this.ensureClientConfigured();
@@ -593,7 +624,7 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
     );
   }
 
-  doc(pathString: string): firestore.DocumentReference {
+  doc(pathString: string): PublicDocumentReference {
     validateExactNumberOfArgs('Firestore.doc', arguments, 1);
     validateArgType('Firestore.doc', 'non-empty string', 1, pathString);
     this.ensureClientConfigured();
@@ -604,7 +635,7 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
     );
   }
 
-  collectionGroup(collectionId: string): firestore.Query {
+  collectionGroup(collectionId: string): PublicQuery {
     validateExactNumberOfArgs('Firestore.collectionGroup', arguments, 1);
     validateArgType(
       'Firestore.collectionGroup',
@@ -628,7 +659,7 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
   }
 
   runTransaction<T>(
-    updateFunction: (transaction: firestore.Transaction) => Promise<T>
+    updateFunction: (transaction: PublicTransaction) => Promise<T>
   ): Promise<T> {
     validateExactNumberOfArgs('Firestore.runTransaction', arguments, 1);
     validateArgType('Firestore.runTransaction', 'function', 1, updateFunction);
@@ -639,13 +670,13 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
     );
   }
 
-  batch(): firestore.WriteBatch {
+  batch(): PublicWriteBatch {
     this.ensureClientConfigured();
 
     return new WriteBatch(this);
   }
 
-  static get logLevel(): firestore.LogLevel {
+  static get logLevel(): PublicLogLevel {
     switch (getLogLevel()) {
       case LogLevel.DEBUG:
         return 'debug';
@@ -665,7 +696,7 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
     }
   }
 
-  static setLogLevel(level: firestore.LogLevel): void {
+  static setLogLevel(level: PublicLogLevel): void {
     validateExactNumberOfArgs('Firestore.setLogLevel', arguments, 1);
     validateStringEnum(
       'setLogLevel',
@@ -681,20 +712,25 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
   _areTimestampsInSnapshotsEnabled(): boolean {
     return this._settings.timestampsInSnapshots;
   }
+
+  // Visible for testing.
+  _getSettings(): PublicSettings {
+    return this._settings;
+  }
 }
 
 /**
  * A reference to a transaction.
  */
-export class Transaction implements firestore.Transaction {
+export class Transaction implements PublicTransaction {
   constructor(
     private _firestore: Firestore,
     private _transaction: InternalTransaction
   ) {}
 
   get<T>(
-    documentRef: firestore.DocumentReference<T>
-  ): Promise<firestore.DocumentSnapshot<T>> {
+    documentRef: PublicDocumentReference<T>
+  ): Promise<PublicDocumentSnapshot<T>> {
     validateExactNumberOfArgs('Transaction.get', arguments, 1);
     const ref = validateReference(
       'Transaction.get',
@@ -737,13 +773,13 @@ export class Transaction implements firestore.Transaction {
   set<T>(
     documentRef: DocumentReference<T>,
     data: Partial<T>,
-    options: firestore.SetOptions
+    options: SetOptions
   ): Transaction;
   set<T>(documentRef: DocumentReference<T>, data: T): Transaction;
   set<T>(
-    documentRef: firestore.DocumentReference<T>,
+    documentRef: PublicDocumentReference<T>,
     value: T | Partial<T>,
-    options?: firestore.SetOptions
+    options?: SetOptions
   ): Transaction {
     validateBetweenNumberOfArgs('Transaction.set', arguments, 2, 3);
     const ref = validateReference(
@@ -770,18 +806,18 @@ export class Transaction implements firestore.Transaction {
   }
 
   update(
-    documentRef: firestore.DocumentReference<unknown>,
-    value: firestore.UpdateData
+    documentRef: PublicDocumentReference<unknown>,
+    value: UpdateData
   ): Transaction;
   update(
-    documentRef: firestore.DocumentReference<unknown>,
+    documentRef: PublicDocumentReference<unknown>,
     field: string | ExternalFieldPath,
     value: unknown,
     ...moreFieldsAndValues: unknown[]
   ): Transaction;
   update(
-    documentRef: firestore.DocumentReference<unknown>,
-    fieldOrUpdateData: string | ExternalFieldPath | firestore.UpdateData,
+    documentRef: PublicDocumentReference<unknown>,
+    fieldOrUpdateData: string | ExternalFieldPath | UpdateData,
     value?: unknown,
     ...moreFieldsAndValues: unknown[]
   ): Transaction {
@@ -825,7 +861,7 @@ export class Transaction implements firestore.Transaction {
     return this;
   }
 
-  delete(documentRef: firestore.DocumentReference<unknown>): Transaction {
+  delete(documentRef: PublicDocumentReference<unknown>): Transaction {
     validateExactNumberOfArgs('Transaction.delete', arguments, 1);
     const ref = validateReference(
       'Transaction.delete',
@@ -837,7 +873,7 @@ export class Transaction implements firestore.Transaction {
   }
 }
 
-export class WriteBatch implements firestore.WriteBatch {
+export class WriteBatch implements PublicWriteBatch {
   private _mutations = [] as Mutation[];
   private _committed = false;
 
@@ -846,13 +882,13 @@ export class WriteBatch implements firestore.WriteBatch {
   set<T>(
     documentRef: DocumentReference<T>,
     data: Partial<T>,
-    options: firestore.SetOptions
+    options: SetOptions
   ): WriteBatch;
   set<T>(documentRef: DocumentReference<T>, data: T): WriteBatch;
   set<T>(
-    documentRef: firestore.DocumentReference<T>,
+    documentRef: PublicDocumentReference<T>,
     value: T | Partial<T>,
-    options?: firestore.SetOptions
+    options?: SetOptions
   ): WriteBatch {
     validateBetweenNumberOfArgs('WriteBatch.set', arguments, 2, 3);
     this.verifyNotCommitted();
@@ -882,18 +918,18 @@ export class WriteBatch implements firestore.WriteBatch {
   }
 
   update(
-    documentRef: firestore.DocumentReference<unknown>,
-    value: firestore.UpdateData
+    documentRef: PublicDocumentReference<unknown>,
+    value: UpdateData
   ): WriteBatch;
   update(
-    documentRef: firestore.DocumentReference<unknown>,
+    documentRef: PublicDocumentReference<unknown>,
     field: string | ExternalFieldPath,
     value: unknown,
     ...moreFieldsAndValues: unknown[]
   ): WriteBatch;
   update(
-    documentRef: firestore.DocumentReference<unknown>,
-    fieldOrUpdateData: string | ExternalFieldPath | firestore.UpdateData,
+    documentRef: PublicDocumentReference<unknown>,
+    fieldOrUpdateData: string | ExternalFieldPath | UpdateData,
     value?: unknown,
     ...moreFieldsAndValues: unknown[]
   ): WriteBatch {
@@ -941,7 +977,7 @@ export class WriteBatch implements firestore.WriteBatch {
     return this;
   }
 
-  delete(documentRef: firestore.DocumentReference<unknown>): WriteBatch {
+  delete(documentRef: PublicDocumentReference<unknown>): WriteBatch {
     validateExactNumberOfArgs('WriteBatch.delete', arguments, 1);
     this.verifyNotCommitted();
     const ref = validateReference(
@@ -979,15 +1015,14 @@ export class WriteBatch implements firestore.WriteBatch {
 /**
  * A reference to a particular document in a collection in the database.
  */
-export class DocumentReference<T = firestore.DocumentData>
-  extends DocumentKeyReference<T>
-  implements firestore.DocumentReference<T> {
+export class DocumentReference<T = DocumentData> extends DocumentKeyReference<T>
+  implements PublicDocumentReference<T> {
   private _firestoreClient: FirestoreClient;
 
   constructor(
     public _key: DocumentKey,
     readonly firestore: Firestore,
-    readonly _converter: firestore.FirestoreDataConverter<T> | null
+    readonly _converter: FirestoreDataConverter<T> | null
   ) {
     super(firestore._databaseId, _key, _converter);
     this._firestoreClient = this.firestore.ensureClientConfigured();
@@ -996,7 +1031,7 @@ export class DocumentReference<T = firestore.DocumentData>
   static forPath<U>(
     path: ResourcePath,
     firestore: Firestore,
-    converter: firestore.FirestoreDataConverter<U> | null
+    converter: FirestoreDataConverter<U> | null
   ): DocumentReference<U> {
     if (path.length % 2 !== 0) {
       throw new FirestoreError(
@@ -1013,7 +1048,7 @@ export class DocumentReference<T = firestore.DocumentData>
     return this._key.path.lastSegment();
   }
 
-  get parent(): firestore.CollectionReference<T> {
+  get parent(): PublicCollectionReference<T> {
     return new CollectionReference(
       this._key.path.popLast(),
       this.firestore,
@@ -1025,9 +1060,7 @@ export class DocumentReference<T = firestore.DocumentData>
     return this._key.path.canonicalString();
   }
 
-  collection(
-    pathString: string
-  ): firestore.CollectionReference<firestore.DocumentData> {
+  collection(pathString: string): PublicCollectionReference<DocumentData> {
     validateExactNumberOfArgs('DocumentReference.collection', arguments, 1);
     validateArgType(
       'DocumentReference.collection',
@@ -1049,7 +1082,7 @@ export class DocumentReference<T = firestore.DocumentData>
     );
   }
 
-  isEqual(other: firestore.DocumentReference<T>): boolean {
+  isEqual(other: PublicDocumentReference<T>): boolean {
     if (!(other instanceof DocumentReference)) {
       throw invalidClassError('isEqual', 'DocumentReference', 1, other);
     }
@@ -1060,9 +1093,9 @@ export class DocumentReference<T = firestore.DocumentData>
     );
   }
 
-  set(value: Partial<T>, options: firestore.SetOptions): Promise<void>;
+  set(value: Partial<T>, options: SetOptions): Promise<void>;
   set(value: T): Promise<void>;
-  set(value: T | Partial<T>, options?: firestore.SetOptions): Promise<void> {
+  set(value: T | Partial<T>, options?: SetOptions): Promise<void> {
     validateBetweenNumberOfArgs('DocumentReference.set', arguments, 1, 2);
     options = validateSetOptions('DocumentReference.set', options);
     const convertedValue = applyFirestoreDataConverter(
@@ -1083,14 +1116,14 @@ export class DocumentReference<T = firestore.DocumentData>
     );
   }
 
-  update(value: firestore.UpdateData): Promise<void>;
+  update(value: UpdateData): Promise<void>;
   update(
     field: string | ExternalFieldPath,
     value: unknown,
     ...moreFieldsAndValues: unknown[]
   ): Promise<void>;
   update(
-    fieldOrUpdateData: string | ExternalFieldPath | firestore.UpdateData,
+    fieldOrUpdateData: string | ExternalFieldPath | UpdateData,
     value?: unknown,
     ...moreFieldsAndValues: unknown[]
   ): Promise<void> {
@@ -1131,21 +1164,19 @@ export class DocumentReference<T = firestore.DocumentData>
     ]);
   }
 
+  onSnapshot(observer: PartialObserver<PublicDocumentSnapshot<T>>): Unsubscribe;
   onSnapshot(
-    observer: PartialObserver<firestore.DocumentSnapshot<T>>
+    options: SnapshotListenOptions,
+    observer: PartialObserver<PublicDocumentSnapshot<T>>
   ): Unsubscribe;
   onSnapshot(
-    options: firestore.SnapshotListenOptions,
-    observer: PartialObserver<firestore.DocumentSnapshot<T>>
-  ): Unsubscribe;
-  onSnapshot(
-    onNext: NextFn<firestore.DocumentSnapshot<T>>,
+    onNext: NextFn<PublicDocumentSnapshot<T>>,
     onError?: ErrorFn,
     onCompletion?: CompleteFn
   ): Unsubscribe;
   onSnapshot(
-    options: firestore.SnapshotListenOptions,
-    onNext: NextFn<firestore.DocumentSnapshot<T>>,
+    options: SnapshotListenOptions,
+    onNext: NextFn<PublicDocumentSnapshot<T>>,
     onError?: ErrorFn,
     onCompletion?: CompleteFn
   ): Unsubscribe;
@@ -1165,7 +1196,7 @@ export class DocumentReference<T = firestore.DocumentData>
       typeof args[currArg] === 'object' &&
       !isPartialObserver(args[currArg])
     ) {
-      options = args[currArg] as firestore.SnapshotListenOptions;
+      options = args[currArg] as SnapshotListenOptions;
       validateOptionNames('DocumentReference.onSnapshot', options, [
         'includeMetadataChanges'
       ]);
@@ -1184,7 +1215,7 @@ export class DocumentReference<T = firestore.DocumentData>
 
     if (isPartialObserver(args[currArg])) {
       const userObserver = args[currArg] as PartialObserver<
-        firestore.DocumentSnapshot<T>
+        PublicDocumentSnapshot<T>
       >;
       args[currArg] = userObserver.next?.bind(userObserver);
       args[currArg + 1] = userObserver.error?.bind(userObserver);
@@ -1213,7 +1244,7 @@ export class DocumentReference<T = firestore.DocumentData>
     const observer: PartialObserver<ViewSnapshot> = {
       next: snapshot => {
         if (args[currArg]) {
-          (args[currArg] as NextFn<firestore.DocumentSnapshot<T>>)(
+          (args[currArg] as NextFn<PublicDocumentSnapshot<T>>)(
             this._convertToDocSnapshot(snapshot)
           );
         }
@@ -1229,7 +1260,7 @@ export class DocumentReference<T = firestore.DocumentData>
     );
   }
 
-  get(options?: firestore.GetOptions): Promise<firestore.DocumentSnapshot<T>> {
+  get(options?: GetOptions): Promise<PublicDocumentSnapshot<T>> {
     validateBetweenNumberOfArgs('DocumentReference.get', arguments, 0, 1);
     validateGetOptions('DocumentReference.get', options);
 
@@ -1256,8 +1287,8 @@ export class DocumentReference<T = firestore.DocumentData>
   }
 
   withConverter<U>(
-    converter: firestore.FirestoreDataConverter<U>
-  ): firestore.DocumentReference<U> {
+    converter: FirestoreDataConverter<U>
+  ): PublicDocumentReference<U> {
     return new DocumentReference<U>(this._key, this.firestore, converter);
   }
 
@@ -1283,13 +1314,13 @@ export class DocumentReference<T = firestore.DocumentData>
   }
 }
 
-export class SnapshotMetadata implements firestore.SnapshotMetadata {
+export class SnapshotMetadata implements PublicSnapshotMetadata {
   constructor(
     readonly hasPendingWrites: boolean,
     readonly fromCache: boolean
   ) {}
 
-  isEqual(other: firestore.SnapshotMetadata): boolean {
+  isEqual(other: PublicSnapshotMetadata): boolean {
     return (
       this.hasPendingWrites === other.hasPendingWrites &&
       this.fromCache === other.fromCache
@@ -1301,20 +1332,20 @@ export class SnapshotMetadata implements firestore.SnapshotMetadata {
  * Options interface that can be provided to configure the deserialization of
  * DocumentSnapshots.
  */
-export interface SnapshotOptions extends firestore.SnapshotOptions {}
+export interface SnapshotOptions extends PublicSnapshotOptions {}
 
-export class DocumentSnapshot<T = firestore.DocumentData>
-  implements firestore.DocumentSnapshot<T> {
+export class DocumentSnapshot<T = DocumentData>
+  implements PublicDocumentSnapshot<T> {
   constructor(
     private _firestore: Firestore,
     private _key: DocumentKey,
     public _document: Document | null,
     private _fromCache: boolean,
     private _hasPendingWrites: boolean,
-    private readonly _converter: firestore.FirestoreDataConverter<T> | null
+    private readonly _converter: FirestoreDataConverter<T> | null
   ) {}
 
-  data(options?: firestore.SnapshotOptions): T | undefined {
+  data(options?: PublicSnapshotOptions): T | undefined {
     validateBetweenNumberOfArgs('DocumentSnapshot.data', arguments, 0, 1);
     options = validateSnapshotOptions('DocumentSnapshot.data', options);
     if (!this._document) {
@@ -1347,7 +1378,7 @@ export class DocumentSnapshot<T = firestore.DocumentData>
 
   get(
     fieldPath: string | ExternalFieldPath,
-    options?: firestore.SnapshotOptions
+    options?: PublicSnapshotOptions
   ): unknown {
     validateBetweenNumberOfArgs('DocumentSnapshot.get', arguments, 1, 2);
     options = validateSnapshotOptions('DocumentSnapshot.get', options);
@@ -1374,7 +1405,7 @@ export class DocumentSnapshot<T = firestore.DocumentData>
     return this._key.path.lastSegment();
   }
 
-  get ref(): firestore.DocumentReference<T> {
+  get ref(): PublicDocumentReference<T> {
     return new DocumentReference<T>(
       this._key,
       this._firestore,
@@ -1386,11 +1417,11 @@ export class DocumentSnapshot<T = firestore.DocumentData>
     return this._document !== null;
   }
 
-  get metadata(): firestore.SnapshotMetadata {
+  get metadata(): PublicSnapshotMetadata {
     return new SnapshotMetadata(this._hasPendingWrites, this._fromCache);
   }
 
-  isEqual(other: firestore.DocumentSnapshot<T>): boolean {
+  isEqual(other: PublicDocumentSnapshot<T>): boolean {
     if (!(other instanceof DocumentSnapshot)) {
       throw invalidClassError('isEqual', 'DocumentSnapshot', 1, other);
     }
@@ -1406,9 +1437,8 @@ export class DocumentSnapshot<T = firestore.DocumentData>
   }
 }
 
-export class QueryDocumentSnapshot<T = firestore.DocumentData>
-  extends DocumentSnapshot<T>
-  implements firestore.QueryDocumentSnapshot<T> {
+export class QueryDocumentSnapshot<T = DocumentData> extends DocumentSnapshot<T>
+  implements PublicQueryDocumentSnapshot<T> {
   data(options?: SnapshotOptions): T {
     const data = super.data(options);
     debugAssert(
@@ -1428,7 +1458,7 @@ export function newQueryFilter(
   op: Operator,
   value: unknown
 ): FieldFilter {
-  let fieldValue: api.Value;
+  let fieldValue: ProtoValue;
   if (fieldPath.isKeyField()) {
     if (op === Operator.ARRAY_CONTAINS || op === Operator.ARRAY_CONTAINS_ANY) {
       throw new FirestoreError(
@@ -1438,8 +1468,8 @@ export function newQueryFilter(
       );
     } else if (op === Operator.IN || op === Operator.NOT_IN) {
       validateDisjunctiveFilterElements(value, op);
-      const referenceList: api.Value[] = [];
-      for (const arrayValue of value as api.Value[]) {
+      const referenceList: ProtoValue[] = [];
+      for (const arrayValue of value as ProtoValue[]) {
         referenceList.push(parseDocumentIdValue(databaseId, query, arrayValue));
       }
       fieldValue = { arrayValue: { values: referenceList } };
@@ -1516,7 +1546,7 @@ export function newQueryBoundFromDocument(
     );
   }
 
-  const components: api.Value[] = [];
+  const components: ProtoValue[] = [];
 
   // Because people expect to continue/end a query at the exact document
   // provided, we need to use the implicit sort order rather than the explicit
@@ -1577,7 +1607,7 @@ export function newQueryBoundFromFields(
     );
   }
 
-  const components: api.Value[] = [];
+  const components: ProtoValue[] = [];
   for (let i = 0; i < values.length; i++) {
     const rawValue = values[i];
     const orderByComponent = orderBy[i];
@@ -1627,7 +1657,7 @@ function parseDocumentIdValue(
   databaseId: DatabaseId,
   query: InternalQuery,
   documentIdValue: unknown
-): api.Value {
+): ProtoValue {
   if (typeof documentIdValue === 'string') {
     if (documentIdValue === '') {
       throw new FirestoreError(
@@ -1828,18 +1858,18 @@ export function validateHasExplicitOrderByForLimitToLast(
   }
 }
 
-export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
+export class Query<T = DocumentData> implements PublicQuery<T> {
   constructor(
     public _query: InternalQuery,
     readonly firestore: Firestore,
-    protected readonly _converter: firestore.FirestoreDataConverter<T> | null
+    protected readonly _converter: FirestoreDataConverter<T> | null
   ) {}
 
   where(
     field: string | ExternalFieldPath,
-    opStr: firestore.WhereFilterOp,
+    opStr: WhereFilterOp,
     value: unknown
-  ): firestore.Query<T> {
+  ): PublicQuery<T> {
     validateExactNumberOfArgs('Query.where', arguments, 3);
     validateDefined('Query.where', 3, value);
 
@@ -1881,8 +1911,8 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
 
   orderBy(
     field: string | ExternalFieldPath,
-    directionStr?: firestore.OrderByDirection
-  ): firestore.Query<T> {
+    directionStr?: OrderByDirection
+  ): PublicQuery<T> {
     validateBetweenNumberOfArgs('Query.orderBy', arguments, 1, 2);
     validateOptionalArgType(
       'Query.orderBy',
@@ -1911,7 +1941,7 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
     );
   }
 
-  limit(n: number): firestore.Query<T> {
+  limit(n: number): PublicQuery<T> {
     validateExactNumberOfArgs('Query.limit', arguments, 1);
     validateArgType('Query.limit', 'number', 1, n);
     validatePositiveNumber('Query.limit', 1, n);
@@ -1922,7 +1952,7 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
     );
   }
 
-  limitToLast(n: number): firestore.Query<T> {
+  limitToLast(n: number): PublicQuery<T> {
     validateExactNumberOfArgs('Query.limitToLast', arguments, 1);
     validateArgType('Query.limitToLast', 'number', 1, n);
     validatePositiveNumber('Query.limitToLast', 1, n);
@@ -1934,9 +1964,9 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
   }
 
   startAt(
-    docOrField: unknown | firestore.DocumentSnapshot<unknown>,
+    docOrField: unknown | PublicDocumentSnapshot<unknown>,
     ...fields: unknown[]
-  ): firestore.Query<T> {
+  ): PublicQuery<T> {
     validateAtLeastNumberOfArgs('Query.startAt', arguments, 1);
     const bound = this.boundFromDocOrFields(
       'Query.startAt',
@@ -1952,9 +1982,9 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
   }
 
   startAfter(
-    docOrField: unknown | firestore.DocumentSnapshot<unknown>,
+    docOrField: unknown | PublicDocumentSnapshot<unknown>,
     ...fields: unknown[]
-  ): firestore.Query<T> {
+  ): PublicQuery<T> {
     validateAtLeastNumberOfArgs('Query.startAfter', arguments, 1);
     const bound = this.boundFromDocOrFields(
       'Query.startAfter',
@@ -1970,9 +2000,9 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
   }
 
   endBefore(
-    docOrField: unknown | firestore.DocumentSnapshot<unknown>,
+    docOrField: unknown | PublicDocumentSnapshot<unknown>,
     ...fields: unknown[]
-  ): firestore.Query<T> {
+  ): PublicQuery<T> {
     validateAtLeastNumberOfArgs('Query.endBefore', arguments, 1);
     const bound = this.boundFromDocOrFields(
       'Query.endBefore',
@@ -1988,9 +2018,9 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
   }
 
   endAt(
-    docOrField: unknown | firestore.DocumentSnapshot<unknown>,
+    docOrField: unknown | PublicDocumentSnapshot<unknown>,
     ...fields: unknown[]
-  ): firestore.Query<T> {
+  ): PublicQuery<T> {
     validateAtLeastNumberOfArgs('Query.endAt', arguments, 1);
     const bound = this.boundFromDocOrFields(
       'Query.endAt',
@@ -2005,7 +2035,7 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
     );
   }
 
-  isEqual(other: firestore.Query<T>): boolean {
+  isEqual(other: PublicQuery<T>): boolean {
     if (!(other instanceof Query)) {
       throw invalidClassError('isEqual', 'Query', 1, other);
     }
@@ -2016,16 +2046,14 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
     );
   }
 
-  withConverter<U>(
-    converter: firestore.FirestoreDataConverter<U>
-  ): firestore.Query<U> {
+  withConverter<U>(converter: FirestoreDataConverter<U>): PublicQuery<U> {
     return new Query<U>(this._query, this.firestore, converter);
   }
 
   /** Helper function to create a bound from a document or fields */
   private boundFromDocOrFields(
     methodName: string,
-    docOrField: unknown | firestore.DocumentSnapshot<T>,
+    docOrField: unknown | PublicDocumentSnapshot<T>,
     fields: unknown[],
     before: boolean
   ): Bound {
@@ -2052,21 +2080,19 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
     }
   }
 
+  onSnapshot(observer: PartialObserver<PublicQuerySnapshot<T>>): Unsubscribe;
   onSnapshot(
-    observer: PartialObserver<firestore.QuerySnapshot<T>>
+    options: SnapshotListenOptions,
+    observer: PartialObserver<PublicQuerySnapshot<T>>
   ): Unsubscribe;
   onSnapshot(
-    options: firestore.SnapshotListenOptions,
-    observer: PartialObserver<firestore.QuerySnapshot<T>>
-  ): Unsubscribe;
-  onSnapshot(
-    onNext: NextFn<firestore.QuerySnapshot<T>>,
+    onNext: NextFn<PublicQuerySnapshot<T>>,
     onError?: ErrorFn,
     onCompletion?: CompleteFn
   ): Unsubscribe;
   onSnapshot(
-    options: firestore.SnapshotListenOptions,
-    onNext: NextFn<firestore.QuerySnapshot<T>>,
+    options: SnapshotListenOptions,
+    onNext: NextFn<PublicQuerySnapshot<T>>,
     onError?: ErrorFn,
     onCompletion?: CompleteFn
   ): Unsubscribe;
@@ -2079,7 +2105,7 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
       typeof args[currArg] === 'object' &&
       !isPartialObserver(args[currArg])
     ) {
-      options = args[currArg] as firestore.SnapshotListenOptions;
+      options = args[currArg] as SnapshotListenOptions;
       validateOptionNames('Query.onSnapshot', options, [
         'includeMetadataChanges'
       ]);
@@ -2094,7 +2120,7 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
 
     if (isPartialObserver(args[currArg])) {
       const userObserver = args[currArg] as PartialObserver<
-        firestore.QuerySnapshot<T>
+        PublicQuerySnapshot<T>
       >;
       args[currArg] = userObserver.next?.bind(userObserver);
       args[currArg + 1] = userObserver.error?.bind(userObserver);
@@ -2118,7 +2144,7 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
     const observer: PartialObserver<ViewSnapshot> = {
       next: snapshot => {
         if (args[currArg]) {
-          (args[currArg] as NextFn<firestore.QuerySnapshot<T>>)(
+          (args[currArg] as NextFn<PublicQuerySnapshot<T>>)(
             new QuerySnapshot(
               this.firestore,
               this._query,
@@ -2137,7 +2163,7 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
     return firestoreClient.listen(this._query, options, observer);
   }
 
-  get(options?: firestore.GetOptions): Promise<firestore.QuerySnapshot<T>> {
+  get(options?: GetOptions): Promise<PublicQuerySnapshot<T>> {
     validateBetweenNumberOfArgs('Query.get', arguments, 0, 1);
     validateGetOptions('Query.get', options);
     validateHasExplicitOrderByForLimitToLast(this._query);
@@ -2153,18 +2179,17 @@ export class Query<T = firestore.DocumentData> implements firestore.Query<T> {
   }
 }
 
-export class QuerySnapshot<T = firestore.DocumentData>
-  implements firestore.QuerySnapshot<T> {
-  private _cachedChanges: Array<firestore.DocumentChange<T>> | null = null;
+export class QuerySnapshot<T = DocumentData> implements PublicQuerySnapshot<T> {
+  private _cachedChanges: Array<PublicDocumentChange<T>> | null = null;
   private _cachedChangesIncludeMetadataChanges: boolean | null = null;
 
-  readonly metadata: firestore.SnapshotMetadata;
+  readonly metadata: PublicSnapshotMetadata;
 
   constructor(
     private readonly _firestore: Firestore,
     private readonly _originalQuery: InternalQuery,
     private readonly _snapshot: ViewSnapshot,
-    private readonly _converter: firestore.FirestoreDataConverter<T> | null
+    private readonly _converter: FirestoreDataConverter<T> | null
   ) {
     this.metadata = new SnapshotMetadata(
       _snapshot.hasPendingWrites,
@@ -2172,8 +2197,8 @@ export class QuerySnapshot<T = firestore.DocumentData>
     );
   }
 
-  get docs(): Array<firestore.QueryDocumentSnapshot<T>> {
-    const result: Array<firestore.QueryDocumentSnapshot<T>> = [];
+  get docs(): Array<PublicQueryDocumentSnapshot<T>> {
+    const result: Array<PublicQueryDocumentSnapshot<T>> = [];
     this.forEach(doc => result.push(doc));
     return result;
   }
@@ -2187,7 +2212,7 @@ export class QuerySnapshot<T = firestore.DocumentData>
   }
 
   forEach(
-    callback: (result: firestore.QueryDocumentSnapshot<T>) => void,
+    callback: (result: PublicQueryDocumentSnapshot<T>) => void,
     thisArg?: unknown
   ): void {
     validateBetweenNumberOfArgs('QuerySnapshot.forEach', arguments, 1, 2);
@@ -2204,13 +2229,11 @@ export class QuerySnapshot<T = firestore.DocumentData>
     });
   }
 
-  get query(): firestore.Query<T> {
+  get query(): PublicQuery<T> {
     return new Query(this._originalQuery, this._firestore, this._converter);
   }
 
-  docChanges(
-    options?: firestore.SnapshotListenOptions
-  ): Array<firestore.DocumentChange<T>> {
+  docChanges(options?: SnapshotListenOptions): Array<PublicDocumentChange<T>> {
     if (options) {
       validateOptionNames('QuerySnapshot.docChanges', options, [
         'includeMetadataChanges'
@@ -2251,7 +2274,7 @@ export class QuerySnapshot<T = firestore.DocumentData>
   }
 
   /** Check the equality. The call can be very expensive. */
-  isEqual(other: firestore.QuerySnapshot<T>): boolean {
+  isEqual(other: PublicQuerySnapshot<T>): boolean {
     if (!(other instanceof QuerySnapshot)) {
       throw invalidClassError('isEqual', 'QuerySnapshot', 1, other);
     }
@@ -2280,12 +2303,12 @@ export class QuerySnapshot<T = firestore.DocumentData>
   }
 }
 
-export class CollectionReference<T = firestore.DocumentData> extends Query<T>
-  implements firestore.CollectionReference<T> {
+export class CollectionReference<T = DocumentData> extends Query<T>
+  implements PublicCollectionReference<T> {
   constructor(
     readonly _path: ResourcePath,
     firestore: Firestore,
-    _converter: firestore.FirestoreDataConverter<T> | null
+    _converter: FirestoreDataConverter<T> | null
   ) {
     super(newQueryForPath(_path), firestore, _converter);
     if (_path.length % 2 !== 1) {
@@ -2302,12 +2325,12 @@ export class CollectionReference<T = firestore.DocumentData> extends Query<T>
     return this._query.path.lastSegment();
   }
 
-  get parent(): firestore.DocumentReference<firestore.DocumentData> | null {
+  get parent(): PublicDocumentReference<DocumentData> | null {
     const parentPath = this._query.path.popLast();
     if (parentPath.isEmpty()) {
       return null;
     } else {
-      return new DocumentReference<firestore.DocumentData>(
+      return new DocumentReference<DocumentData>(
         new DocumentKey(parentPath),
         this.firestore,
         /* converter= */ null
@@ -2319,7 +2342,7 @@ export class CollectionReference<T = firestore.DocumentData> extends Query<T>
     return this._query.path.canonicalString();
   }
 
-  doc(pathString?: string): firestore.DocumentReference<T> {
+  doc(pathString?: string): PublicDocumentReference<T> {
     validateBetweenNumberOfArgs('CollectionReference.doc', arguments, 0, 1);
     // We allow omission of 'pathString' but explicitly prohibit passing in both
     // 'undefined' and 'null'.
@@ -2340,7 +2363,7 @@ export class CollectionReference<T = firestore.DocumentData> extends Query<T>
     );
   }
 
-  add(value: T): Promise<firestore.DocumentReference<T>> {
+  add(value: T): Promise<PublicDocumentReference<T>> {
     validateExactNumberOfArgs('CollectionReference.add', arguments, 1);
     const convertedValue = this._converter
       ? this._converter.toFirestore(value)
@@ -2351,16 +2374,16 @@ export class CollectionReference<T = firestore.DocumentData> extends Query<T>
   }
 
   withConverter<U>(
-    converter: firestore.FirestoreDataConverter<U>
-  ): firestore.CollectionReference<U> {
+    converter: FirestoreDataConverter<U>
+  ): PublicCollectionReference<U> {
     return new CollectionReference<U>(this._path, this.firestore, converter);
   }
 }
 
 function validateSetOptions(
   methodName: string,
-  options: firestore.SetOptions | undefined
-): firestore.SetOptions {
+  options: SetOptions | undefined
+): SetOptions {
   if (options === undefined) {
     return {
       merge: false
@@ -2391,8 +2414,8 @@ function validateSetOptions(
 
 function validateSnapshotOptions(
   methodName: string,
-  options: firestore.SnapshotOptions | undefined
-): firestore.SnapshotOptions {
+  options: PublicSnapshotOptions | undefined
+): PublicSnapshotOptions {
   if (options === undefined) {
     return {};
   }
@@ -2410,7 +2433,7 @@ function validateSnapshotOptions(
 
 function validateGetOptions(
   methodName: string,
-  options: firestore.GetOptions | undefined
+  options: GetOptions | undefined
 ): void {
   validateOptionalArgType(methodName, 'object', 1, options);
   if (options) {
@@ -2427,7 +2450,7 @@ function validateGetOptions(
 
 function validateReference<T>(
   methodName: string,
-  documentRef: firestore.DocumentReference<T>,
+  documentRef: PublicDocumentReference<T>,
   firestore: Firestore
 ): DocumentKeyReference<T> {
   if (!(documentRef instanceof DocumentKeyReference)) {
@@ -2443,14 +2466,14 @@ function validateReference<T>(
 }
 
 /**
- * Calculates the array of firestore.DocumentChange's for a given ViewSnapshot.
+ * Calculates the array of DocumentChanges for a given ViewSnapshot.
  *
  * Exported for testing.
  *
  * @param snapshot The ViewSnapshot that represents the expected state.
  * @param includeMetadataChanges Whether to include metadata changes.
  * @param converter A factory function that returns a QueryDocumentSnapshot.
- * @return An objecyt that matches the firestore.DocumentChange API.
+ * @return An object that matches the DocumentChange API.
  */
 export function changesFromSnapshot<DocSnap>(
   snapshot: ViewSnapshot,
@@ -2461,7 +2484,7 @@ export function changesFromSnapshot<DocSnap>(
     hasPendingWrite: boolean
   ) => DocSnap
 ): Array<{
-  type: firestore.DocumentChangeType;
+  type: DocumentChangeType;
   doc: DocSnap;
   oldIndex: number;
   newIndex: number;
@@ -2487,7 +2510,7 @@ export function changesFromSnapshot<DocSnap>(
       );
       lastDoc = change.doc;
       return {
-        type: 'added' as firestore.DocumentChangeType,
+        type: 'added' as DocumentChangeType,
         doc,
         oldIndex: -1,
         newIndex: index++
@@ -2523,7 +2546,7 @@ export function changesFromSnapshot<DocSnap>(
   }
 }
 
-function resultChangeType(type: ChangeType): firestore.DocumentChangeType {
+function resultChangeType(type: ChangeType): DocumentChangeType {
   switch (type) {
     case ChangeType.Added:
       return 'added';
@@ -2549,8 +2572,8 @@ function resultChangeType(type: ChangeType): firestore.DocumentChangeType {
 export function applyFirestoreDataConverter<T>(
   converter: UntypedFirestoreDataConverter<T> | null,
   value: T,
-  options?: firestore.SetOptions
-): firestore.DocumentData {
+  options?: SetOptions
+): DocumentData {
   let convertedValue;
   if (converter) {
     if (options && (options.merge || options.mergeFields)) {
@@ -2562,7 +2585,7 @@ export function applyFirestoreDataConverter<T>(
       convertedValue = converter.toFirestore(value);
     }
   } else {
-    convertedValue = value as firestore.DocumentData;
+    convertedValue = value as DocumentData;
   }
   return convertedValue;
 }
