@@ -15,7 +15,10 @@
  * limitations under the License.
  */
 
-import * as api from '../protos/firestore_proto_api';
+import {
+  Value as ProtoValue,
+  MapValue as ProtoMapValue
+} from '../protos/firestore_proto_api';
 
 import { SnapshotVersion } from '../core/snapshot_version';
 import { fail } from '../util/assert';
@@ -67,7 +70,7 @@ export class Document extends MaybeDocument {
     this.hasCommittedMutations = !!options.hasCommittedMutations;
   }
 
-  field(path: FieldPath): api.Value | null {
+  field(path: FieldPath): ProtoValue | null {
     return this.objectValue.field(path);
   }
 
@@ -75,7 +78,7 @@ export class Document extends MaybeDocument {
     return this.objectValue;
   }
 
-  toProto(): { mapValue: api.MapValue } {
+  toProto(): { mapValue: ProtoMapValue } {
     return this.objectValue.proto;
   }
 
