@@ -15,13 +15,16 @@
  * limitations under the License.
  */
 
-import * as firestore from '@firebase/firestore-types';
+import {
+  FirestoreError as PublicFirestoreError,
+  FirestoreErrorCode as PublicFirestoreErrorCode
+} from '@firebase/firestore-types';
 
 /**
  * Error Codes describing the different ways Firestore can fail. These come
  * directly from GRPC.
  */
-export type Code = firestore.FirestoreErrorCode;
+export type Code = PublicFirestoreErrorCode;
 
 export const Code = {
   // Causes are copied from:
@@ -158,7 +161,7 @@ export const Code = {
  * so we define our own compatible error class (with a `name` of 'FirebaseError'
  * and compatible `code` and `message` fields.)
  */
-export class FirestoreError extends Error implements firestore.FirestoreError {
+export class FirestoreError extends Error implements PublicFirestoreError {
   name = 'FirebaseError';
   stack?: string;
 
