@@ -1266,7 +1266,7 @@ apiDescribe('Database', (persistence: boolean) => {
     });
   });
 
-  it('calling terminate mutiple times should proceed', async () => {
+  it('calling terminate multiple times should proceed', async () => {
     await withTestDoc(persistence, async docRef => {
       const firestore = docRef.firestore;
       await firestore.terminate();
