@@ -19,7 +19,6 @@ import { stub } from 'sinon';
 import { expect } from 'chai';
 import { Api, setupApi } from './api_service';
 import '../../test/setup';
-
 describe('Firebase Performance > api_service', () => {
   const PAGE_URL = 'http://www.test.com/abcd?a=2';
   const PERFORMANCE_ENTRY: PerformanceEntry = {
@@ -47,6 +46,7 @@ describe('Firebase Performance > api_service', () => {
     ]);
     // This is to make sure the test page is not changed by changing the href of location object.
     mockWindow.location = { ...self.location, href: PAGE_URL };
+
     setupApi(mockWindow);
     api = Api.getInstance();
   });
